requirements.txt
openai
sentence-transformers
faiss-cpu
numpy
python-dotenv 
chromadb